data/requirements.txt
# A-Lab Experiments - Complete Python Dependencies
# ===================================================
# Install with: pip install -r requirements.txt
# Or use automated setup: bash setup_migration_env.sh

# Core Database & Migration
pymongo>=4.0.0              # MongoDB client
psycopg2-binary>=2.9.0      # PostgreSQL client
python-dateutil>=2.8.0      # Date parsing
python-dotenv>=1.0          # Environment variables (.env files)

# Dashboard & Visualization (Plotly Dash)
dash==2.18.1                # Dash framework
dash-bootstrap-components==1.6.0  # Bootstrap UI components
plotly==5.24.1              # Interactive plots

# Data Processing
pandas==2.2.3               # DataFrames and analysis
numpy==2.1.3                # Numerical operations
pyarrow>=14.0.0             # Parquet file support
rich>=13.0.0                # Beautiful terminal output
tqdm>=4.65.0                # Progress bars

# Product Pipeline
pydantic>=2.0               # Data validation
PyYAML>=6.0                 # Configuration files
inquirer>=3.0               # Interactive CLI

# Migration Pipeline
tqdm>=4.65.0                # Progress bars
# mpcontribs-client>=5.0      # MPContribs API
# mp-api>=0.35                # Materials Project API
boto3>=1.40.0               # AWS S3 uploads for MPContribs OpenData
